feat(plugin): adapter plugin -- presto and postgresql
