+++ configure/etc/conf/plugins.properties
@@ -32,6 +32,8 @@ datacap-plugin-neo4j=plugin/datacap-plugin-neo4j/pom.xml
 datacap-plugin-oceanbase=plugin/datacap-plugin-oceanbase/pom.xml
 datacap-plugin-oracle=plugin/datacap-plugin-oracle/pom.xml
 datacap-plugin-phoenix=plugin/datacap-plugin-phoenix/pom.xml
+datacap-plugin-postgresql=plugin/datacap-plugin-postgresql/pom.xml
+datacap-plugin-presto=plugin/datacap-plugin-presto/pom.xml
 datacap-convert-txt=convert/datacap-convert-txt/pom.xml
 datacap-convert-json=convert/datacap-convert-json/pom.xml
 datacap-convert-xml=convert/datacap-convert-xml/pom.xml
