--- plugin/datacap-jdbc-postgresql/src/main/java/io/edurt/datacap/plugin/jdbc/postgresql/PostgreSQLAdapter.java
@@ -1,15 +1,0 @@
-package io.edurt.datacap.plugin.jdbc.postgresql;
-
-import io.edurt.datacap.spi.adapter.JdbcAdapter;
-import io.edurt.datacap.spi.connection.JdbcConnection;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-public class PostgreSQLAdapter
-        extends JdbcAdapter
-{
-    public PostgreSQLAdapter(JdbcConnection jdbcConnection)
-    {
-        super(jdbcConnection);
-    }
-}
