--- plugin/datacap-jdbc-postgresql/src/main/java/io/edurt/datacap/plugin/jdbc/postgresql/PostgreSQLConnection.java
@@ -1,16 +1,0 @@
-package io.edurt.datacap.plugin.jdbc.postgresql;
-
-import io.edurt.datacap.spi.connection.JdbcConfigure;
-import io.edurt.datacap.spi.connection.JdbcConnection;
-import io.edurt.datacap.spi.model.Response;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-public class PostgreSQLConnection
-        extends JdbcConnection
-{
-    public PostgreSQLConnection(JdbcConfigure jdbcConfigure, Response response)
-    {
-        super(jdbcConfigure, response);
-    }
-}
