--- plugin/datacap-jdbc-presto/src/main/java/io/edurt/datacap/plugin/jdbc/presto/PrestoAdapter.java
@@ -1,13 +1,0 @@
-package io.edurt.datacap.plugin.jdbc.presto;
-
-import io.edurt.datacap.spi.adapter.JdbcAdapter;
-import io.edurt.datacap.spi.connection.JdbcConnection;
-
-public class PrestoAdapter
-        extends JdbcAdapter
-{
-    public PrestoAdapter(JdbcConnection jdbcConnection)
-    {
-        super(jdbcConnection);
-    }
-}
