--- plugin/datacap-jdbc-presto/src/test/java/io/edurt/datacap/plugin/jdbc/presto/PrestoPluginTest.java
@@ -1,45 +1,0 @@
-package io.edurt.datacap.plugin.jdbc.presto;
-
-import com.google.inject.Guice;
-import com.google.inject.Injector;
-import com.google.inject.Key;
-import com.google.inject.TypeLiteral;
-import io.edurt.datacap.spi.PluginService;
-import io.edurt.datacap.spi.model.Configure;
-import org.junit.Before;
-import org.junit.Test;
-
-import java.util.Optional;
-import java.util.Set;
-
-public class PrestoPluginTest
-{
-    private Injector injector;
-    private Configure configure;
-
-    @Before
-    public void before()
-    {
-        injector = Guice.createInjector(new PrestoPluginModule());
-        configure = new Configure();
-        configure.setHost("12.0.0.1");
-        configure.setPort(8080);
-        configure.setUsername(Optional.of("default"));
-        configure.setPassword(Optional.of(""));
-    }
-
-    @Test
-    public void test()
-    {
-        Set<PluginService> plugins = injector.getInstance(Key.get(new TypeLiteral<Set<PluginService>>() {}));
-        Optional<PluginService> pluginOptional = plugins.stream()
-                .filter(v -> v.name().equalsIgnoreCase("Presto"))
-                .findFirst();
-        if (pluginOptional.isPresent()) {
-            PluginService plugin = pluginOptional.get();
-            plugin.connect(configure);
-            System.out.println(plugin.execute("SHOW CATALOGS"));
-            plugin.destroy();
-        }
-    }
-}
