+++ plugin/datacap-plugin-postgresql/src/main/java/io/edurt/datacap/plugin/PostgreSQLPlugin.java
@@ -1,0 +1,5 @@
+package io.edurt.datacap.plugin;
+
+public class PostgreSQLPlugin
+        extends Plugin
+{}
