+++ plugin/datacap-plugin-postgresql/src/main/java/io/edurt/datacap/plugin/PostgreSQLService.java
@@ -1,0 +1,19 @@
+package io.edurt.datacap.plugin;
+
+import io.edurt.datacap.spi.PluginService;
+
+public class PostgreSQLService
+        implements PluginService
+{
+    @Override
+    public String driver()
+    {
+        return "org.postgresql.Driver";
+    }
+
+    @Override
+    public String connectType()
+    {
+        return "postgresql";
+    }
+}
