+++ plugin/datacap-plugin-postgresql/src/main/resources/META-INF/services/io.edurt.datacap.plugin.Service
@@ -1,0 +1,1 @@
+io.edurt.datacap.plugin.PostgreSQLService
