+++ plugin/datacap-plugin-presto/src/main/java/io/edurt/datacap/plugin/PrestoPlugin.java
@@ -1,0 +1,5 @@
+package io.edurt.datacap.plugin;
+
+public class PrestoPlugin
+        extends Plugin
+{}
